Move calculate_inputs_statistics to helpers module
